Move scanner out of executor
