+++ src/coprocessor/dag/executor/mod.rs
@@ -33,7 +33,6 @@ mod aggregation;
 mod index_scan;
 mod limit;
 mod scan;
-pub mod scanner;
 mod selection;
 mod table_scan;
 mod topn;
@@ -46,7 +45,6 @@ pub use self::index_scan::IndexScanExecutor;
 pub use self::limit::LimitExecutor;
 pub use self::metrics::*;
 pub use self::scan::ScanExecutor;
-pub use self::scanner::{ScanOn, Scanner};
 pub use self::selection::SelectionExecutor;
 pub use self::table_scan::TableScanExecutor;
 pub use self::topn::TopNExecutor;
