+++ src/coprocessor/dag/scanner.rs
@@ -15,7 +15,7 @@ use kvproto::coprocessor::KeyRange;
 
 use crate::coprocessor::codec::table::truncate_as_row_key;
 use crate::coprocessor::util;
-use crate::storage::txn::Result;
+use crate::coprocessor::Result;
 use crate::storage::{Key, Scanner as KvScanner, Statistics, Store, Value};
 use crate::util::{escape, set_panic_mark};
 
@@ -35,7 +35,7 @@ fn create_range_scanner<S: Store>(
 ) -> Result<S::Scanner> {
     let lower_bound = Some(Key::from_raw(range.get_start()));
     let upper_bound = Some(Key::from_raw(range.get_end()));
-    store.scanner(desc, key_only, lower_bound, upper_bound)
+    Ok(store.scanner(desc, key_only, lower_bound, upper_bound)?)
 }
 
 // `Scanner` is a helper struct to wrap all common scan operations
@@ -178,7 +178,7 @@ pub mod tests {
     use kvproto::kvrpcpb::IsolationLevel;
     use tipb::schema::ColumnInfo;
 
-    use super::super::tests::{get_range, new_col_info, TestStore};
+    use super::super::executor::tests::{get_range, new_col_info, TestStore};
     use crate::coprocessor::codec::datum::{self, Datum};
     use crate::coprocessor::codec::table;
     use crate::coprocessor::util;
